refactor(services): :recycle: Remove dead code from the ftu data services
