+++ libs/services/src/index.ts
@@ -1,4 +1,3 @@
-export { Iri, Url } from './lib/shared/common.model';
 export {
   Biomarker,
   Cell,
@@ -7,19 +6,17 @@ export {
   DataFileReference,
   IllustrationMappingItem,
   SourceReference,
+  Tissue,
+  TissueLibrary,
 } from './lib/ftu-data/ftu-data.model';
-export { Tissue, TissueLibrary } from './lib/tissue-library/tissue-library.model';
+export { Iri, Url } from './lib/shared/common.model';
 
-export * from './lib/contact/contact.service';
 export * from './lib/contact/contact.mock';
-export * from './lib/ftu-data/ftu-data.service';
+export * from './lib/contact/contact.service';
 export * from './lib/ftu-data/ftu-data.impl';
 export * from './lib/ftu-data/ftu-data.mock';
+export * from './lib/ftu-data/ftu-data.service';
 export * from './lib/service.module';
 export * from './lib/shared/common.model';
-export * from './lib/tissue-library/tissue-library.service';
-export * from './lib/tissue-library/tissue-library.mock';
-export * from './lib/tissue-ftu/tissue-ftu.service';
-export * from './lib/tissue-ftu/tissue-ftu.mock';
 
 export * as FtuDataSchemas from './lib/ftu-data/ftu-data.model';
